requirements.txt

# Requirements for TMvisDB

pymongo[srv]
pandas
py3Dmol
stmol
requests
streamlit-aggrid
